Say hello to side panels
The list of buffers (on the left) and users (on the right, when
on a channel) have been moved side panels that can be opened by
swiping side ways. This makes it convenient to quickly peek the
list of buffers whether there are any new/highlighted messages,
whilst the real screen estate is still reserved to the textual
content.